requirements.txt
ibm-cos-sdk
python-dotenv
pandas
matplotlib
seaborn
